ai_engine/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.9.0
appnope==0.1.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
blis==0.7.9
brotlipy==0.7.0
catalogue==2.0.8
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.3
comm==0.1.2
confection==0.0.4
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.115.11
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
flask-restx==1.1.0
frozenlist==1.3.3
gdown==4.6.4
h11==0.14.0
idna==3.10
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
JPype1==1.4.1
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.3.0
ko-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_lg-3.5.0/ko_core_news_lg-3.5.0-py3-none-any.whl#sha256=8251bb7a7ea6b44dfd4deb07b8e5815ab816331449a82cd27af62b6deffb598c
konlpy==0.6.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
numpy==1.24.2
openai==0.27.4
packaging==23.0
pandas==2.0.0
parso==0.8.3
Paste==3.5.2
pathy==0.10.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==3.1.1
preshed==3.0.8
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.10.1
PyMySQL==1.0.3
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyupbit==0.2.34
pyzmq==25.0.1
requests==2.32.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.1
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.46.1
thinc==8.1.9
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
uvicorn==0.34.0
waitress==2.1.2
wasabi==1.1.1
wcwidth==0.2.6
websockets==14.1
Werkzeug==2.2.3
wincertstore==0.2
yarl==1.8.2
zstandard==0.18.0
